examples/stt_deepgram_transcription/pyproject.toml
[project]
name = "getstream-stt-deepgram-transcription-example"
version = "0.1.0"
description = "Example project showing how to transcribe a call using STT with Deepgram"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "getstream>=1.4.0",
    "deepgram-sdk>=3.11.0",
    "python-dotenv>=1.0.0",
    "aiortc>=1.10.1",
    "numpy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
